% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_snps.R
\name{cluster_snps_dt}
\alias{cluster_snps_dt}
\title{Cluster SNPs Based on Genomic Distance and P-value (Data.Table Version)}
\usage{
cluster_snps_dt(
  data,
  chr_col = "chr",
  pos_col = "pos",
  rsid_col = "rsid",
  pvalue_col = "pvalue",
  gene_col = "gene",
  pvalue_threshold = 0.05,
  distance_threshold = 1e+06
)
}
\arguments{
\item{data}{A data.frame containing SNP information.}

\item{chr_col}{Character. Name of the chromosome column. Default is "chr".}

\item{pos_col}{Character. Name of the position column. Default is "pos".}

\item{rsid_col}{Character. Name of the SNP ID column. Default is "rsid".}

\item{pvalue_col}{Character. Name of the p-value column. Default is "pvalue".}

\item{gene_col}{Character. Name of the gene column (optional). Default is "gene".}

\item{pvalue_threshold}{Numeric. Maximum p-value for SNPs to be considered. Default is 0.05.}

\item{distance_threshold}{Numeric. Maximum distance (bp) between SNPs to belong to the same cluster. Default is 1e6.}
}
\value{
A data.table with cluster assignments.
}
\description{
This function clusters SNPs by chromosome and genomic distance using an efficient
algorithm that avoids large distance matrices.
}
