% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_pair_plot.R
\name{construct_plot_title}
\alias{construct_plot_title}
\title{Helper function to generate a formatted plot title}
\usage{
construct_plot_title(
  plot_title_prefix,
  n_cases,
  n_controls,
  lambda,
  total_snps_in_study,
  add_date_to_title,
  gwas_data,
  font_family = "Arial Unicode MS"
)
}
\arguments{
\item{plot_title_prefix}{The prefix for the plot title (e.g., study name).}

\item{n_cases}{The number of cases in the study.}

\item{n_controls}{The number of controls in the study.}

\item{lambda}{The genomic inflation factor (lambda GC).}

\item{total_snps_in_study}{The total number of SNPs in the study.}

\item{add_date_to_title}{A logical flag to append the current date to the title.}

\item{gwas_data}{The GWAS dataset.}

\item{font_family}{The font family to use for the title.}
}
\value{
A formatted string for the plot title.
}
\description{
Helper function to generate a formatted plot title
}
