% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_pair_plot.R
\name{create_mirrored_manhattan_plot}
\alias{create_mirrored_manhattan_plot}
\title{Create a Pair of Inverted Manhattan Plots with Advanced Features}
\usage{
create_mirrored_manhattan_plot(
  gwas_data1,
  gwas_data2,
  chr_col = "chr",
  bp_col = "pos",
  p_col = "pvalue",
  rsid_col = NULL,
  rsids_to_remove = NULL,
  gene_col = "gene",
  group_col = NULL,
  loci_to_label = NULL,
  genes_to_label = NULL,
  custom_gene_colors = NULL,
  n_cases1,
  n_controls1,
  n_cases2,
  n_controls2,
  plot_title1 = "Dataset 1",
  plot_title2 = "Dataset 2",
  lambda1 = NULL,
  lambda2 = NULL,
  total_snps_in_study1 = NULL,
  total_snps_in_study2 = NULL,
  add_date_to_title = FALSE,
  plot_pval_threshold = 1,
  sig_lines = c(red = 5e-08, blue = 1e-05),
  label_threshold_colors = c(red = 5e-08, orange = 1e-06, darkblue = 1e-05),
  y_axis_squish_threshold = 30,
  gene_label_size = 3.5,
  label_orientation = "horizontal",
  label_alpha = 0.7,
  output_folder = "Inverted_Manhattan_Plots",
  file_name_prefix = "inverted_manhattan",
  font_family = "Arial Unicode MS",
  output_width = 18,
  output_height = 12
)
}
\arguments{
\item{gwas_data1}{The first GWAS dataset.}

\item{gwas_data2}{The second GWAS dataset.}

\item{chr_col}{Column name for chromosome.}

\item{bp_col}{Column name for base pair position.}

\item{p_col}{Column name for p-value.}

\item{rsid_col}{Optional: Column name for SNP identifiers (e.g., RSIDs) for filtering.}

\item{rsids_to_remove}{Optional: A vector of SNP IDs to remove from both datasets before analysis.}

\item{gene_col}{Column name for gene labels.}

\item{group_col}{Column for grouping labels. Defaults to gene_col.}

\item{n_cases1}{Number of cases in dataset 1.}

\item{n_controls1}{Number of controls in dataset 1.}

\item{n_cases2}{Number of cases in dataset 2.}

\item{n_controls2}{Number of controls in dataset 2.}

\item{plot_title1}{Title for the top plot.}

\item{plot_title2}{Title for the bottom plot.}

\item{lambda1}{Optional lambda for dataset 1.}

\item{lambda2}{Optional lambda for dataset 2.}

\item{label_orientation}{Orientation of gene labels ("vertical" or "horizontal").}

\item{label_alpha}{Transparency of gene label backgrounds (0-1).}

\item{output_width}{Width of the output plot in inches.}

\item{output_height}{Height of the output plot in inches.}

\item{...}{other parameters}
}
\description{
(Version 18: Synced features with single plot function)
}
