% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload.R
\name{.onLoad}
\alias{.onLoad}
\title{Package startup message}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library path}

\item{pkgname}{Package name}
}
\description{
This function is called when the package is loaded and displays
information about the package creators.
}
