% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_pair_plot.R
\name{generate_manhattan_subplot}
\alias{generate_manhattan_subplot}
\title{Helper function to generate a single Manhattan plot (normal or inverted)}
\usage{
generate_manhattan_subplot(
  plot_data,
  axis_data,
  plot_title,
  sig_lines,
  label_threshold_colors,
  y_axis_squish_threshold,
  gene_label_size,
  gene_col,
  group_col,
  loci_to_label,
  genes_to_label,
  custom_gene_colors,
  inverted,
  font_family = "Arial Unicode MS",
  label_orientation = "horizontal",
  label_alpha = 0.7
)
}
\description{
(Internal function for create_mirrored_manhattan_plot)
}
