% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gwasdataseta.R
\docType{data}
\name{gwasdataseta}
\alias{gwasdataseta}
\title{Example GWAS dataset}
\format{
A data frame with 21,946 rows and 5 variables:
\describe{
  \item{chr}{Chromosome number (as character).}
  \item{pos}{Base pair position of the SNP (integer).}
  \item{pvalue}{P-value for association (numeric).}
  \item{gene}{Associated gene name (character).}
  \item{rsid}{Reference SNP ID (character).}
}
}
\source{
Simulated dataset for package example purposes.
}
\usage{
gwasdataseta
}
\description{
A simulated GWAS dataset containing SNP information, chromosomal positions,
p-values, and associated gene names. This dataset is intended for use
as an example in demonstrations and package vignettes.
}
\keyword{datasets}
